requirements.txt
accelerate==1.11.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.54.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
argcomplete==3.6.2
asyncpg==0.30.0
attrs==25.3.0
backoff==2.2.1
beautifulsoup4==4.14.2
blis==1.3.0
boto3==1.38.41
botocore==1.38.41
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.2.1
cloudpathlib==0.23.0
cohere==5.15.0
colorama==0.4.6
colorlog==6.10.1
confection==0.1.5
contourpy==1.3.3
cycler==0.12.1
cymem==2.0.11
dill==0.4.0
diskcache==5.6.3
distro==1.9.0
docling==2.60.0
docling-core==2.50.0
docling-ibm-models==3.10.2
docling-parse==4.7.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
et_xmlfile==2.0.0
eval_type_backport==0.2.2
Faker==37.12.0
fasta2a==0.3.2
fastapi==0.115.13
fastavro==1.11.1
filelock==3.18.0
filetype==1.2.0
fonttools==4.60.1
frozenlist==1.7.0
fsspec==2025.5.1
google-auth==2.40.3
google-genai==1.43.0
graphiti==0.1.13
graphiti-core==0.22.0
greenlet==3.2.4
griffe==1.7.3
groq==0.28.0
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.36.0
idna==3.11
importlib_metadata==8.7.0
iniconfig==2.1.0
Jinja2==3.1.6
jiter==0.11.1
jmespath==1.0.1
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
langchain-core==0.3.79
langcodes==3.5.0
langgraph==0.6.10
langgraph-checkpoint==2.1.2
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.9
langsmith==0.4.37
language_data==1.3.0
latex2mathml==3.78.1
logfire-api==3.21.1
lxml==6.0.2
marisa-trie==1.3.1
markdown-it-py==3.0.0
marko==2.2.1
MarkupSafe==3.0.3
matplotlib==3.10.7
mcp==1.9.4
mdurl==0.1.2
mistralai==1.8.2
mpire==2.10.2
mpmath==1.3.0
multidict==6.5.0
multiprocess==0.70.18
murmurhash==1.0.13
neo4j==6.0.2
networkx==3.5
numpy==2.3.4
omegaconf==2.3.0
openai==2.7.1
opencv-python==4.12.0.88
openpyxl==3.1.5
opentelemetry-api==1.34.1
orjson==3.11.3
ormsgpack==1.11.0
packaging==25.0
pandas==2.3.3
pillow==11.3.0
pluggy==1.6.0
polyfactory==2.22.3
posthog==6.8.0
preshed==3.0.10
prompt_toolkit==3.0.51
propcache==0.3.2
psutil==7.1.3
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclipper==1.3.0.post6
pydantic==2.12.3
pydantic-ai==0.3.2
pydantic-ai-slim==0.3.2
pydantic-evals==0.3.2
pydantic-graph==0.3.2
pydantic-settings==2.10.0
pydantic_core==2.41.4
Pygments==2.19.2
pygraphviz==1.14
pylatexenc==2.10
pyparsing==3.2.5
pypdfium2==4.30.0
pytest==8.4.1
pytest-asyncio==1.0.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-multipart==0.0.20
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.2
rapidocr==3.4.2
referencing==0.37.0
regex==2025.10.23
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.28.0
rsa==4.9.1
rtree==1.4.1
s3transfer==0.13.0
safetensors==0.6.2
scipy==1.16.3
seaborn==0.13.2
semchunk==2.2.2
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
smart_open==7.3.1
sniffio==1.3.1
soupsieve==2.8
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.44
srsly==2.5.1
sse-starlette==2.3.6
starlette==0.46.2
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
thinc==8.3.6
tokenizers==0.22.1
torch==2.9.0+cpu
torchvision==0.24.0+cpu
tqdm==4.67.1
transformers==4.57.1
typer==0.19.2
types-requests==2.32.4.20250611
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uuid==1.30
uvicorn==0.34.3
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
websockets==15.0.1
wrapt==1.17.3
xlsxwriter==3.2.9
xxhash==3.6.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.25.0
